doublesRun/requirements.txt
Django==1.11.10
django-cors-headers==2.2.0
djangorestframework==3.7.7
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
